5/30/21 Update
Added column for best AOE ammotype.
Fixed bug in AOE damage module that was defaulting a lot of values to 0 when they should not have.
Added a check to compare ammo damage multipliers and if damagepershot is the same to compare the based on multipliers in order to capture the highest damage ammo in those scenarios.