Upgraded version 2.0.0 now including a GUI
This major update integrates the Python TKinter package into the tool allowing a much better user experience (could be better laid out I'm sure but what am I? A UX designer??). TKinter now allows the tool to have a GUI and allows the user to select their install filepath to analyze instead of a hard-coded variable.

The tool has also been packaged in a onefile executable allowing anyone to run the tool standalone without needing to install Python or any library dependencies. There are two versions of this executable, one that includes a visible console to use the logging checkbox and the other hides the console for a cleaner experience. Most users should opt for the noconsole version unless they want a peek under the hood during the process but beware, when logging is on the tool runtime drastically increases.

I have also finally included a logging toggle variable, within the GUI this will be visible as a checkbox. Disabling logging when not needed reduces tool runtime down to around 4 seconds from my previous 500+ seconds with all logging on.